+++ projects/deja-js/component/v2/numeric-stepper/numeric-stepper.component.spec.ts
@@ -10,9 +10,14 @@ import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { FormBuilder, FormGroup, ReactiveFormsModule } from '@angular/forms';
 import { By } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { ControlsOf } from '@deja-js/component/core';
 
 import { DejaNumericStepperModule } from '.';
 import { DejaNumericStepperComponent } from './numeric-stepper.component';
+
+interface NumberForm {
+    numberValue: number;
+}
 
 @Component({
     selector: 'DejaNumericStepperContainerComponent',
@@ -21,7 +26,7 @@ import { DejaNumericStepperComponent } from './numeric-stepper.component';
                 <deja-numeric-stepper (increment)="numberValue = numberValue + step" (decrement)="numberValue = numberValue - step"></deja-numeric-stepper></mat-form-field>`
 })
 class DejaNumericStepperContainerComponent {
-    public numberForm: FormGroup;
+    public numberForm: FormGroup<ControlsOf<NumberForm>>;
     public numberValue = 30;
     public step = 10;