+++ src/app/v2/monaco-editor/monaco-editor-demo.module.ts
@@ -15,8 +15,8 @@ import { MatSelectModule } from '@angular/material/select';
 import { MatTabsModule } from '@angular/material/tabs';
 import { MatToolbarModule } from '@angular/material/toolbar';
 import { RouterModule, Routes } from '@angular/router';
-import { DejaSplitterModule } from '@deja-js/component/splitter';
 import { MonacoEditorModule } from '@deja-js/component/v2/monaco-editor';
+import { DejaSplitterModule } from '@deja-js/component/v2/splitter';
 
 import { DejaMarkdownModule } from '../../../component/markdown/index';
 import { MonacoEditorDemoComponent } from './monaco-editor-demo.component';